Move proposal sign function to hook
